Fixed casting issues after crc32 changes for mac.

+++ mz_strm_lzma.c
@@ -402,9 +402,9 @@ void *mz_stream_lzma_get_interface(void)
     return (void *)&mz_stream_lzma_vtbl;
 }
 
-static int32_t mz_stream_lzma_crc32(int32_t value, const void *buf, int32_t size)
+static int64_t mz_stream_lzma_crc32(int64_t value, const void *buf, int32_t size)
 {
-    return lzma_crc32(buf, size, value);
+    return (int32_t)lzma_crc32(buf, size, (int32_t)value);
 }
 
 void *mz_stream_lzma_get_crc32_table(void)
